requirements.txt
selenium==4.3.0
webdriver-manager==3.8.3
python-dotenv==0.21.0
